Extract calculations to a separate package
