+++ apps/api/package.json
@@ -26,6 +26,7 @@
     "@hapi/joi": "17.1.1",
     "@hapi/vision": "6.0.1",
     "@prisma/client": "2.10.2",
+    "@sklep/calculations": "^1.0.0-alpha.0",
     "bcrypt": "5.0.0",
     "bluebird": "3.7.2",
     "dotenv": "8.2.0",
