+++ apps/api/src/plugins/cart/cartFunctions.ts
@@ -1,4 +1,5 @@
 import type { Request, ResponseToolkit } from '@hapi/hapi';
+import { calculateCartTotals } from '@sklep/calculations';
 
 import type { Awaited } from '../../types';
 
@@ -144,27 +145,6 @@ export function findAllCarts(request: Request) {
   return request.server.app.db.cart.findMany({
     select: cartSelect,
   });
-}
-
-export function calculateCartTotals(cart: CartFromDB) {
-  return cart.cartProducts.reduce(
-    (acc, cartProduct) => {
-      const regularSum = cartProduct.product.regularPrice * cartProduct.quantity;
-      const discountSum =
-        (cartProduct.product.discountPrice ?? cartProduct.product.regularPrice) *
-        cartProduct.quantity;
-
-      acc.regularSubTotal += Math.trunc(regularSum);
-      acc.discountSubTotal += Math.trunc(discountSum);
-      acc.totalQuantity += cartProduct.quantity;
-      return acc;
-    },
-    {
-      regularSubTotal: 0,
-      discountSubTotal: 0,
-      totalQuantity: 0,
-    },
-  );
 }
 
 export const cartModelToResponse = (cart: CartFromDB) => {
