+++ apps/calculations/.eslintrc
@@ -1,0 +1,9 @@
+{
+  "root": false,
+  "parserOptions": {
+    "project": "tsconfig.json"
+  },
+  "rules": {
+    "import/no-default-export": "error"
+  }
+}
