+++ apps/calculations/jest-setup.ts
@@ -1,0 +1,3 @@
+process.on('unhandledRejection', (err) => {
+  fail(err);
+});
