+++ apps/calculations/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  testEnvironment: 'node',
+  transform: {
+    '^.+\\.(t|j)sx?$': ['@swc-node/jest'],
+  },
+  setupFiles: ['./jest-setup.ts'],
+  setupFilesAfterEnv: [],
+};
