+++ apps/calculations/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "@sklep/calculations",
+  "version": "1.0.0-alpha.0",
+  "private": true,
+  "scripts": {
+    "dev": "jest --watchAll",
+    "test": "jest",
+    "test_:ci": "jest"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "@swc-node/jest": "1.0.1",
+    "@tsconfig/node12": "1.0.7",
+    "@types/faker": "5.1.4",
+    "faker": "5.1.0",
+    "jest": "26.6.3",
+    "typescript": "4.0.5"
+  }
+}
