+++ package.json
@@ -2,11 +2,12 @@
   "name": "sklep.typeofweb",
   "private": "true",
   "scripts": {
-    "dev": "yarn --frozen-lockfile && lerna bootstrap && lerna run dev --parallel",
+    "dev": "yarn --frozen-lockfile && lerna run dev --parallel",
     "eslint": "yarn lerna run eslint --stream",
     "test": "yarn lerna run test --stream",
     "test:ci": "yarn lerna run test_:ci --stream",
-    "tsc": "yarn lerna run tsc --stream"
+    "tsc": "yarn lerna run tsc --stream",
+    "postinstall": "lerna bootstrap"
   },
   "workspaces": {
     "packages": [
